+++ changelog.md
@@ -9,6 +9,7 @@
 * Re-added "Actual redstone component"
 * Fixed Forge's items on top being rotated wrongly
 * Decreased entity reach of hammers by 0.25
+* Increased repair chance of Solarium equipment
 * Fixed Forge and Ensorceller not syncing items on world load
 * Hopefully fixed client-side crash when interacting with Wandering Trader on a server